+++ db.js
@@ -1,0 +1,11 @@
+const sqlite3 = require('sqlite3').verbose();
+
+let db = new sqlite3.Database('./db/appsploit.db', sqlite3.OPEN_READWRITE, (err) => {
+  if (err) {
+    console.error(err.message);
+  } else {
+    console.log('Connected to the appsploit database.');
+  }
+});
+
+module.exports = db;
