+++ index.js
@@ -10,6 +10,20 @@ const session = require('express-session');
 const sqlite3 = require('sqlite3').verbose();
 const app = express();
 
+// Database object
+var db = require('./db')
+
+//
+// Require modular routing files
+//
+const routeInstructions = require("./routes/routeInstructions");
+const routeSecureTasks = require("./routes/secureRoutes/routeSecureTasks");
+const routeToggleSecure = require("./routes/routeToggleSecure");
+const routePostTask = require("./routes/secureRoutes/routePostTask");
+const routeXSSTasks = require("./routes/XSSRoutes/routeXSSTasks");
+const routeXSSPostTask = require("./routes/XSSRoutes/routeXSSPostTask");
+const routeToggleTask = require("./routes/routeToggleTask");
+const routeDeleteTask = require("./routes/routeDeleteTask");
 
 //
 // Configuration
@@ -31,144 +45,29 @@ app.use((req, res, next) => {
     next();
 });
 
+/*******************************************************************************
+* Set routes here
+*
+* Controller functions
+********************************************************************************/
 
-// 
-// Connect to database
-// 
-let db = new sqlite3.Database('./db/appsploit.db', sqlite3.OPEN_READWRITE, (err) => {
-  if (err) {
-    console.error(err.message);
-  } else {
-    console.log('Connected to the appsploit database.');
-  }
-});
+// Secure routes
+app.get('/', (req, res, next) => routeSecureTasks(req, res, next));
+app.post('/', (req, res, next) => routePostTask(req, res, next));
 
+// XSS routes
+app.get('/xss', (req, res, next) => routeXSSTasks(req, res, next));
+app.post('/xss', (req, res, next) => routeXSSPostTask(req, res, next));
 
-//
-// Application Endpoints
-//
-app.get('/togglesecure', function(req, res, next) {
+// Static pages & general routes
+app.get('/instructions', (req, res, next) => routeInstructions(req, res, next));
+app.get('/togglesecure', (req, res, next) => routeToggleSecure(req, res, next));
 
-    if (req.session.secure) {
-        req.session.secure = false;
-        res.json({"secure": false})
-    } else {
-        req.session.secure = true;
-        res.json({"secure": true})
-    }
-});
-
-// 
-// Static pages
-// 
-app.get('/instructions', function(req, res, next){
-    let context = {
-        vulnerability: "Select a vulnerability",
-    };
-    res.render('instructions', context);
-});
-
-// 
-// Default Todo Behavior
-// 
-app.get('/', function(req, res, next){
-    let context = {
-        vulnerability: "Select a vulnerability",
-        endpoint: req.originalUrl,
-        exploit_card: 'default_card'
-    };
-
-    db.all("select * from todo", [], (err, rows) => {
-        context.tasks = rows
-        res.render('secure_tasks', context);
-    })
-});
-
-app.post('/', function(req, res, next) {
-    
-    data = [req.body.desc, 0, 1]
-
-    sql = "insert into todo(task_description, task_complete, user_id) values(?,?,?)"
-    db.run(sql, data, function(err){
-        if (err) {
-            console.error(err.message)
-        } 
-    })
-    res.redirect(req.originalUrl)
-})
-
-// 
-// XSS
-// 
-app.get('/xss', function(req, res, next){
-    let context = {
-        vulnerability: "Cross-site scripting (XSS)",
-        endpoint: req.originalUrl,
-        exploit_card: 'xss_card'
-    };
-
-    db.all("select * from todo", [], (err, rows) => {
-        context.tasks = rows
-
-        if(req.session.secure) {
-            res.render('secure_tasks', context);
-        } else{
-            res.render('xss_tasks', context);
-        }
-    })
-});
-
-app.post('/xss', function(req, res, next) {
-    
-    data = [req.body.desc, 0, 1]
-
-    sql = "insert into todo(task_description, task_complete, user_id) values(?,?,?)"
-    db.run(sql, data, function(err){
-        if (err) {
-            console.error(err.message)
-        } 
-    })
-    res.redirect(req.originalUrl)
-})
-
-
-// 
 // Task Endpoints
-// 
-app.get('/task/:task_id', function(req, res, next) {
+app.get('/task/:task_id', (req, res, next) => routeToggleTask(req, res, next));
+app.get('/task/:task_id/delete', (req, res, next) => routeDeleteTask(req, res, next));
 
-    var complete = 0;
-    if (req.query.complete === "true") {
-        complete = 1
-    }
-    data = [complete, req.params.task_id]
-    sql = "update todo set task_complete = ? where task_id = ?"
-
-    db.run(sql, data, function(err){
-        if (err) {
-            console.error(err.message)
-        } else {
-            res.json({"complete": req.query.complete})
-        }
-    })
-
-})
-
-app.get('/task/:task_id/delete', function(req, res, next) {
-    data = [req.params.task_id]
-    sql = "delete from todo where task_id = ?"
-    db.run(sql, data, function(err){
-        if (err) {
-            console.error(err.message)
-        } else {
-            res.sendStatus(200);
-        }
-    });
-});
-
-//
 // Error Handlers
-//
 app.use(function(req, res){
     res.status(404);
     res.render('404');
@@ -179,6 +78,10 @@ app.use(function(err, req, res, next){
     res.status(500);
     res.render('500');
 });
+
+/*******************************************************************************
+ * End controller functions
+ ******************************************************************************/
 
 
 //
