+++ routes/routeInstructions.js
@@ -1,0 +1,8 @@
+const routeInstructions = (req, res, next) => {
+    let context = {
+        vulnerability: "Select a vulnerability",
+    };
+    res.render('instructions', context);
+};
+
+module.exports = routeInstructions;
